Add files via upload
Monitoring the air quality in a smart city. 
Using an Arduino Uno board along with various sensors to detect harmful gases and measure temperature and humidity, I have developed a system that collects environmental data through electrical components. This system not only gathers the data but also sends it to an IoT platform for visualization via a MATLAB extension. The platform allows users to monitor pollution levels in real-time, helping to identify areas that are best avoided due to high pollution.